Metadata Review - FLORT-D and FLORD-G
I have completed the changes that were documented in the CGSN Metadata Review log for these two intstruments. Any changes I made are doccumented in the notes column and are in italics. A lot of the changes were related to the significant digits of scattering scale factor. Everytime this was the case, I made the correct change too the number and also changed the formatting of the cell to show the appropriate number of significant digits. Hopefully the formatting change survives the merge, though I don't think it will. When the reviewers open the csvs again, the cell may go back to the old formatting but if you click on the shell it should show the actual value.