ADACPA-N Review Complete
Our AUV adcps have not returned to the vendor for any work since we received them and we do not seem to have any paperwork from the vendor or hydroid that would give us a clue which day we received them.

There were 2 cal csvs per adcp SN in the system already, both likely for deployments. Since these contain constants only, I deleted one for each SN.  I then changed the date on the cal csv for ADCPA-N 22076 so that it matched with 22077, both of which are now 5/7/2015, which is one day prior to our first AUV deployment with ingestible data.