Create CGINS-ADCPAM-55979__20200709.csv
This creates a "calibration csv" for a newly acquired glider DVL transducer. The transducer in DVL bay 541 was recently replaced and since we now track DVL transducer SNs instead of DVL bay numbers, we need to add a calibration csv for this new transducer. Since the transducers do not actually get calibrated, we use acquisition or vendor service dates for the date to be used in the cal csv.

The transducer swap can be tracked in: $/Project_Files/Records/Platform_Records/Gliders/Coastal_glider/DVL/CSR_5878_DVL-541_2020-07-09.pdf

And glider 387's FCP: $/Project_Files/Records/Platform_Records/3304-00150-00129-A.zip/387_Pioneer_15a_FCP.cap CTRL-F for the text "PS0" to show the output of when the DVL is queried for its transducer SN